requirements.txt
lxml==4.8.0
psycopg2-binary==2.9.3
requests==2.27.1
